backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.3.3
reportlab==4.0.7
python-decouple==3.8
